photobooth/requirements.txt
numpy
opencv-python
Pillow
gpiozero
pycups
requests
PyQt6
picamera
picamera2
gphoto2
gphoto2-cffi
Flask